+++ MPChartExample/src/com/xxmassdeveloper/mpchartexample/CombinedChartActivity.java
@@ -69,19 +69,19 @@ public class CombinedChartActivity extends DemoBase {
 
         YAxis rightAxis = mChart.getAxisRight();
         rightAxis.setDrawGridLines(false);
-        rightAxis.setAxisMinimum(0f); // this replaces setStartAtZero(true)
+        rightAxis.setAxisMinimum(0); // this replaces setStartAtZero(true)
 
         YAxis leftAxis = mChart.getAxisLeft();
         leftAxis.setDrawGridLines(false);
-        leftAxis.setAxisMinimum(0f); // this replaces setStartAtZero(true)
+        leftAxis.setAxisMinimum(0); // this replaces setStartAtZero(true)
 
         XAxis xAxis = mChart.getXAxis();
         xAxis.setPosition(XAxisPosition.BOTH_SIDED);
-        xAxis.setAxisMinimum(0f);
-        xAxis.setGranularity(1f);
+        xAxis.setAxisMinimum(0);
+        xAxis.setGranularity(1);
         xAxis.setValueFormatter(new IAxisValueFormatter() {
             @Override
-            public String getFormattedValue(float value, AxisBase axis) {
+            public String getFormattedValue(double value, AxisBase axis) {
                 return mMonths[(int) value % mMonths.length];
             }
         });
@@ -95,7 +95,7 @@ public class CombinedChartActivity extends DemoBase {
         data.setData(generateCandleData());
         data.setValueTypeface(mTfLight);
 
-        xAxis.setAxisMaximum(data.getXMax() + 0.25f);
+        xAxis.setAxisMaximum(data.getXMax() + 0.25);
 
         mChart.setData(data);
         mChart.invalidate();
@@ -108,17 +108,17 @@ public class CombinedChartActivity extends DemoBase {
         ArrayList<Entry> entries = new ArrayList<Entry>();
 
         for (int index = 0; index < itemcount; index++)
-            entries.add(new Entry(index + 0.5f, getRandom(15, 5)));
+            entries.add(new Entry(index + 0.5, getRandom(15, 5)));
 
         LineDataSet set = new LineDataSet(entries, "Line DataSet");
         set.setColor(Color.rgb(240, 238, 70));
-        set.setLineWidth(2.5f);
+        set.setLineWidth(2.5);
         set.setCircleColor(Color.rgb(240, 238, 70));
-        set.setCircleRadius(5f);
+        set.setCircleRadius(5);
         set.setFillColor(Color.rgb(240, 238, 70));
         set.setMode(LineDataSet.Mode.CUBIC_BEZIER);
         set.setDrawValues(true);
-        set.setValueTextSize(10f);
+        set.setValueTextSize(10);
         set.setValueTextColor(Color.rgb(240, 238, 70));
 
         set.setAxisDependency(YAxis.AxisDependency.LEFT);
@@ -136,25 +136,25 @@ public class CombinedChartActivity extends DemoBase {
             entries1.add(new BarEntry(0, getRandom(25, 25)));
 
             // stacked
-            entries2.add(new BarEntry(0, new float[]{getRandom(13, 12), getRandom(13, 12)}));
+            entries2.add(new BarEntry(0, new double[]{getRandom(13, 12), getRandom(13, 12)}));
         }
 
         BarDataSet set1 = new BarDataSet(entries1, "Bar 1");
         set1.setColor(Color.rgb(60, 220, 78));
         set1.setValueTextColor(Color.rgb(60, 220, 78));
-        set1.setValueTextSize(10f);
+        set1.setValueTextSize(10);
         set1.setAxisDependency(YAxis.AxisDependency.LEFT);
 
         BarDataSet set2 = new BarDataSet(entries2, "");
         set2.setStackLabels(new String[]{"Stack 1", "Stack 2"});
         set2.setColors(new int[]{Color.rgb(61, 165, 255), Color.rgb(23, 197, 255)});
         set2.setValueTextColor(Color.rgb(61, 165, 255));
-        set2.setValueTextSize(10f);
+        set2.setValueTextSize(10);
         set2.setAxisDependency(YAxis.AxisDependency.LEFT);
 
-        float groupSpace = 0.06f;
-        float barSpace = 0.02f; // x2 dataset
-        float barWidth = 0.45f; // x2 dataset
+        double groupSpace = 0.06;
+        double barSpace = 0.02; // x2 dataset
+        double barWidth = 0.45; // x2 dataset
         // (0.45 + 0.02) * 2 + 0.06 = 1.00 -> interval per "group"
 
         BarData d = new BarData(set1, set2);
@@ -172,14 +172,14 @@ public class CombinedChartActivity extends DemoBase {
 
         ArrayList<Entry> entries = new ArrayList<Entry>();
 
-        for (float index = 0; index < itemcount; index += 0.5f)
-            entries.add(new Entry(index + 0.25f, getRandom(10, 55)));
+        for (double index = 0; index < itemcount; index += 0.5)
+            entries.add(new Entry(index + 0.25, getRandom(10, 55)));
 
         ScatterDataSet set = new ScatterDataSet(entries, "Scatter DataSet");
         set.setColors(ColorTemplate.MATERIAL_COLORS);
-        set.setScatterShapeSize(7.5f);
+        set.setScatterShapeSize(7.5);
         set.setDrawValues(false);
-        set.setValueTextSize(10f);
+        set.setValueTextSize(10);
         d.addDataSet(set);
 
         return d;
@@ -192,13 +192,13 @@ public class CombinedChartActivity extends DemoBase {
         ArrayList<CandleEntry> entries = new ArrayList<CandleEntry>();
 
         for (int index = 0; index < itemcount; index += 2)
-            entries.add(new CandleEntry(index + 1f, 90, 70, 85, 75f));
+            entries.add(new CandleEntry(index + 1, 90, 70, 85, 75));
 
         CandleDataSet set = new CandleDataSet(entries, "Candle DataSet");
         set.setDecreasingColor(Color.rgb(142, 150, 175));
         set.setShadowColor(Color.DKGRAY);
-        set.setBarSpace(0.3f);
-        set.setValueTextSize(10f);
+        set.setBarSpace(0.3);
+        set.setValueTextSize(10);
         set.setDrawValues(false);
         d.addDataSet(set);
 
@@ -212,16 +212,16 @@ public class CombinedChartActivity extends DemoBase {
         ArrayList<BubbleEntry> entries = new ArrayList<BubbleEntry>();
 
         for (int index = 0; index < itemcount; index++) {
-            float y = getRandom(10, 105);
-            float size = getRandom(100, 105);
-            entries.add(new BubbleEntry(index + 0.5f, y, size));
+            double y = getRandom(10, 105);
+            double size = getRandom(100, 105);
+            entries.add(new BubbleEntry(index + 0.5, y, size));
         }
 
         BubbleDataSet set = new BubbleDataSet(entries, "Bubble DataSet");
         set.setColors(ColorTemplate.VORDIPLOM_COLORS);
-        set.setValueTextSize(10f);
+        set.setValueTextSize(10);
         set.setValueTextColor(Color.WHITE);
-        set.setHighlightCircleWidth(1.5f);
+        set.setHighlightCircleWidth(1.5);
         set.setDrawValues(true);
         bd.addDataSet(set);
